Move common test functions to own module
